Remove CONFIG reads from Slide.render
Slide.render is called during playback, and we want the CONFIG values from
slide construction.  Slide.finalize's purpose is to lock in those values
from construction time so a slide deck like this can have per-slide
configurations:
```
slide
	config 'pad.top', 10
	title 'hello'

slide
	config 'pad.top', 1
	title 'goodbye'
```

If the values are nil at render time, that means we're missing a call to
Slide.finalize.  It's not something we should fix by moving those CONFIG
reads to Slide.render.

The CONFIG reads were added to make sure that `@transition` isn't nil, but
we also now check `transition` before using it, so that's no longer
needed.